Initialize the interpreter only once. This fixes the crashes I've been having all the time.
Note, in passing, that the recent move of plparrot.c to src, instead of
src/handler, doesn't fix the problem where make installs everything into the
wrong place. PostgreSQL still looks for plparrot.so in $libdir, but the
installer puts it in $libdir/src. This hasn't been failing immediately in my
case, because there's an old plparrot.so already there. That makes debugging
... less simple. :) But that's also a fix for another night.