Give pgTAP a much-needed haircut and forego using pg_prove
It turns out that pgTAP is a bit of a moving target and is really tied to the
version of Postgres that is being used.  We get rid of all functions that
deal with schema-verification and use psql directly instead of using pg_prove,
so as to not have another external dependency that users need to run our tests.

Your welcome.