initial SPI support via NCI
Several new NCI thunks are required for some of these to work.  The
complete list is as follows:

    i
    i    p
    i    pi
    i    ppil
    i    pppil
    i    pppl
    i    til
    i    tipppil
    i    tl
    p    l
    p    p
    p    pl
    p    t
    p    tip
    p    tipi
    p    tppi
    p    tpppi
    p    ttipppii
    v
    v    it
    v    p
    v    pil
    v    tit

and some other functions could not be supported at all, due to structs
being passed.  These functions might later be wrapped, or we could wait
for the libffi NCI refactor.

To add these to Parrot, add the above signatures to
src/nci/extra_thunks.nci, then run

    ./parrot_nci_thunk_gen --core --loader-name=Parrot_nci_load_extra_thunks \
        --no-warn-dups --output=src/extra/nci_thunks.c <src/nci/extra_thunks.nci