Allow Perl 6 Grammars as return values of stored procedures
When defining a global grammar for later use, the return value of the
stored procedure was a Grammar object, which PL/Perl6 did not know how
to convert to a Postgres datatype, so the sausage machine came to a
grinding halt. We catch these and just return a true value instead.

There is currently a bug in Rakudo where Grammar.WHAT is "Code", and
somehow that gets turned into a "Sub", so that was worked around as
well.